# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital card API",
        "description": "This API allows users to create, update, and delete business cards.",
        "version": "1.0.0"
    },
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de l'inscription de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion de l'utilisateur",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Utilisateur déconnecté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cartes-visites": {
            "get": {
                "tags": [
                    "Cartes de visite"
                ],
                "summary": "Récupérer toutes les cartes de visite de l'utilisateur",
                "operationId": "6cd07726eb7196ef1dfa4e0136244e5b",
                "responses": {
                    "200": {
                        "description": "Liste des cartes de visite de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CarteVisite"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cartes de visite"
                ],
                "summary": "Créer une nouvelle carte de visite",
                "operationId": "1f22b9a3780185e9a0a4f0bc63635031",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "tel",
                                    "entreprise",
                                    "titre",
                                    "coordonnees",
                                    "description"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "string"
                                    },
                                    "entreprise": {
                                        "type": "string"
                                    },
                                    "titre": {
                                        "type": "string"
                                    },
                                    "coordonnees": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carte de visite créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "carteVisite": {
                                            "$ref": "#/components/schemas/CarteVisite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la création de la carte de visite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cartes-visites/{id}": {
            "put": {
                "tags": [
                    "Cartes de visite"
                ],
                "summary": "Mettre à jour une carte de visite",
                "operationId": "3f543f51d1e076ab31c31588ae30881d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carte de visite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "string"
                                    },
                                    "entreprise": {
                                        "type": "string"
                                    },
                                    "titre": {
                                        "type": "string"
                                    },
                                    "coordonnees": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Carte de visite mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "carteVisite": {
                                            "$ref": "#/components/schemas/CarteVisite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carte de visite non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cartes de visite"
                ],
                "summary": "Supprimer une carte de visite",
                "operationId": "6453df5faf2c7fbc654bf0fbc9344a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la carte de visite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Carte de visite supprimée avec succès"
                    },
                    "403": {
                        "description": "Non autorisé: Vous n'avez pas la permission de supprimer cette carte de visite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Carte de visite non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CarteVisite": {
                "required": [
                    "id",
                    "nom",
                    "tel",
                    "entreprise",
                    "title",
                    "coordonnees",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "tel": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "entreprise": {
                        "type": "string",
                        "example": "XYZ Company"
                    },
                    "title": {
                        "type": "string",
                        "example": "CEO"
                    },
                    "coordonnees": {
                        "type": "string",
                        "example": "123 Main Street, City, Country"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                    }
                },
                "type": "object"
            }
        }
    }
}